+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### Version 4.18.0 (4th July 2019)
+#### Added
+- Added `trackAdRevenue` method to `Adjust` interface to allow tracking of ad revenue. With this release added support for `MoPub` ad revenue tracking.
+- Added reading of Facebook anonymous ID if available on iOS platform.
+
+#### Native SDKs
+- [iOS@v4.18.0][ios_sdk_v4.18.0]
+- [Android@v4.18.0][android_sdk_v4.18.0]
+
+---
+
 ### Version 4.17.1 (5th June 2019)
 #### Fixed
 - Fixed issue when trying to register Android plugin more than once (https://github.com/adjust/flutter_sdk/issues/7).
@@ -60,5 +71,7 @@
 
 [ios_sdk_v4.17.0]: https://github.com/adjust/ios_sdk/tree/v4.17.0
 [ios_sdk_v4.17.3]: https://github.com/adjust/ios_sdk/tree/v4.17.3
+[ios_sdk_v4.18.0]: https://github.com/adjust/ios_sdk/tree/v4.18.0
 
 [android_sdk_v4.17.0]: https://github.com/adjust/android_sdk/tree/v4.17.0
+[android_sdk_v4.18.0]: https://github.com/adjust/android_sdk/tree/v4.18.0